requirements.txt
pandas>=2.0
python-dotenv>=1.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9